requirements.txt
chembl-webresource-client>=0.10.8
rdkit-pypi>=2022.9.5
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
shap>=0.44.0
optuna>=3.5.0
streamlit>=1.30.0
joblib>=1.3.0
transformers>=4.40.0
torch>=2.0.0
PyTDC>=0.4.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
jupyter>=1.0.0
ipywidgets>=8.0.0
